* commit '044e1b87f5334ac23f14aa8af62a813f5e7fd1c8':
  some fixes
  some changes
  search Api and UI changes

+++ app/src/main/java/com/hexagram/febys/models/api/response/SearchSuggestionPagingListing.kt
@@ -12,7 +12,7 @@ class SearchSuggestionPagingListing(
     @SerializedName("pagingInfo", alternate = ["pagination_info"])
     val pagingInfo: PagingInfo,
     @SerializedName("total_rows")
-    val totalRows: Int,
+    val totalRows: Long,
     val hasNextPage: Boolean,
     val hasPreviousPage: Boolean,
     val nextPage: Int,
